Rules updated and FieldErrorProcessor
The rules have been updated to all work with empty strings/arrays, which historically could not work, this overlaps with how required rule works, so if you NEED something to have a value then it should be required, then the other rules should be adding more specificity to the object.

Also have updated the field error processor to output a contextual error if a rule is not found when validating.