Have fixed issue with array indexes being removed
In cases where arrays have validation on the child elements, there is an issue where the element can be removed but the validation error is then stuck, so this clears down child validation errors before it is revalidated.

This can potentially cause events to trigger a bit more often for array scenarios but for now its not the end of the world.